Split test into separate class
